% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SplitDF}
\alias{SplitDF}
\title{Split dataframe.}
\usage{
SplitDF(df, key_col, val_col)
}
\arguments{
\item{df}{Dataframe.}

\item{key_col}{The columns used as keys.}

\item{val_col}{The columns used as values, this column can contain key: value pairs.}
}
\value{
List of dataframe.
}
\description{
Split dataframe.
}
\examples{
# df <- data.frame(
#   A = c("a", "a", "", "", "", "f"),
#   B = c("a", "b", "", "", "", "f"),
#   C = c("a", "b", "c", "", "", "f"),
#   D = c("a", "b", "c", "d", "", "f"),
#   E = c("a", "b", "c", "d", "e", "f"),
#   G = c("a", "b", "c", "d", "e", "f")
# ) \%>\%
#   t() \%>\%
#   as.data.frame()
# colnames(df) <- c("c1", "c2", "c3", "c4", "c5", "c6")
# SplitDF(df = df, key_col = c("c1", "c2", "c3", "c4", "c5"), val_col = "c6")
}
