% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2yaml.R
\name{df2yaml}
\alias{df2yaml}
\title{Convert dataframe to YAML.}
\usage{
df2yaml(
  df,
  key_col = c("paras", "subcmd"),
  val_col = "values",
  val_sep = ";",
  key_sep = ":",
  out_yaml = NULL
)
}
\arguments{
\item{df}{Dataframe.}

\item{key_col}{The columns used as keys, up to 2.}

\item{val_col}{The columns used as values, this column can contain key: value pairs.}

\item{val_sep}{The separator used to seperate different key:value pairs in \code{val_col}. Default: ";".}

\item{key_sep}{The separator used to seperate key and value. Default: ":".}

\item{out_yaml}{The output yaml file. Default: NULL (df2yaml.yaml).}
}
\value{
NULL (save results to file.)
}
\description{
Convert dataframe to YAML.
}
\examples{

}
